+++ src/spikes/ble_handoff/DashboardBootSwitch.h
@@ -4,5 +4,7 @@ namespace dashboard_boot {
 
 bool switchToReader();
 bool switchToReceiver();
+bool isRunningReader();
+bool isRunningReceiver();
 
 }  // namespace dashboard_boot